Query: Include: Refactor to use tree data structure inside IncludeCompiler.
