+++ src/EFCore/Query/Expressions/Internal/NullConditionalExpression.cs
@@ -140,39 +140,6 @@ namespace Microsoft.EntityFrameworkCore.Query.Expressions.Internal
         }
 
         /// <summary>
-        ///     Tests if this object is considered equal to another.
-        /// </summary>
-        /// <param name="obj"> The object to compare with the current object. </param>
-        /// <returns>
-        ///     true if the objects are considered equal, false if they are not.
-        /// </returns>
-        public override bool Equals(object obj)
-        {
-            if (ReferenceEquals(null, obj))
-            {
-                return false;
-            }
-
-            if (ReferenceEquals(this, obj))
-            {
-                return true;
-            }
-
-            return obj.GetType() == GetType() && Equals((NullConditionalExpression)obj);
-        }
-
-        private bool Equals(NullConditionalExpression other)
-            => Equals(AccessOperation, other.AccessOperation);
-
-        /// <summary>
-        ///     Returns a hash code for this object.
-        /// </summary>
-        /// <returns>
-        ///     A hash code for this object.
-        /// </returns>
-        public override int GetHashCode() => AccessOperation.GetHashCode();
-
-        /// <summary>
         ///     Returns a textual representation of the <see cref="T:System.Linq.Expressions.Expression" />.
         /// </summary>
         /// <returns>
